requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.13
djangorestframework==3.12.4
idna==3.2
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.62.0
urllib3==1.26.6
authy~=2.2.6